Import all enums into the null-group
and make sure that we do not import enums having different names but
pointing to the same value. The problem is that even though we have enum
groups, those groups are not always declared on the command GLenum
parameters. One example is the 'target' of glCopyBufferSubData(). It
should belong to "BufferTargetARB", but is not declared as such.