Resolve declaring class of static method
With the latest change in LWJGL 3.1.7 that a GLnn class extends from
GL(nn-1) class, a static call on GL30.glDrawArrays() is now possible and
the static receiver type of that call will be GL30.
So, we have to manually resolve the actual declaring class of the
method, in order to correctly resolve our own trace and validation
methods.